[cffLib] Add remove_hints() and remove_unused_subroutines() methods
From subset.cff.
